% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bring_files_xml.R
\name{bring_files}
\alias{bring_files}
\title{bring files from the server to another folder}
\usage{
bring_files(
  file_sel = "xml",
  folder_sel = "E:/test",
  country_sel,
  folder_out = "data/xml",
  time_min = "2022-01-01"
)
}
\arguments{
\item{file_sel}{type of files ("xml", "csv",...). "xml" is the default option.}

\item{folder_sel}{folder to look at (not recursively)}

\item{country_sel}{country to look for}

\item{folder_out}{folder where to store files. Default is  "data/xml"}

\item{time_min}{date to the file from where to start}
}
\value{
files are copied, if they exist they are not overwritten
}
\description{
bring files from the server to another folder
}
\examples{
bring_files(country_sel="PL")
}
