% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gva_pyp.R
\name{calculate_gva_pyp}
\alias{calculate_gva_pyp}
\title{Calculate B1G in previous year prices for Matis}
\usage{
calculate_gva_pyp(country_sel, nat_dat, reg_dat, output_dir)
}
\arguments{
\item{country_sel}{Country}

\item{nat_dat}{A dataframe with the national data on GVA at previous year prices. It should have the following columns: country, sto, prices, time_period, obs_value.}

\item{reg_dat}{A dataframe with regional data. It should have the following columns: country, ref_area, NUTS, sto, prices, time_period, obs_value.}

\item{output_dir}{the folder where t create the file}
}
\value{
a csv file
}
\description{
The function creates a csv that can be later on loaded in Matis.
}
\examples{
calculate_gva_pyp("LU",nat_nama,regacc,"vol_GDP/output")
}
