% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NUTS.R
\name{check_NUTS}
\alias{check_NUTS}
\title{Check NUTS additivity}
\usage{
check_NUTS(data, ths_abs = 2, ths_rel = 0.05)
}
\arguments{
\item{data}{a data frame}

\item{ths_abs}{absolute threshold}

\item{ths_rel}{relative threshold in \% (1=1\%)}
}
\value{
a data frame with the data not fulfilling the consistency requirements
}
\description{
Function for checking NUTS additivity. It requires a data frame with at least the following columns
country,ref_area,table_identifier,time_period, sto, accounting_entry, activity, unit_measure, obs_value
}
\examples{
df <- load_csv(
folder = "E:/data/REGACC/csv",
country_sel = c("AT"),
table_sel = "T1002",
time_min = "2021-12-01")

check_NUTS(df)


}
