% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check packages}
\usage{
check_packages()
}
\description{
The function checks that the packages needed are installed and loads them
}
\examples{
check_packages()
}
