% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_anything.R
\name{convert_anything}
\alias{convert_anything}
\title{Convert most common file formats}
\usage{
convert_anything(dir, origin, dest, recursive = FALSE)
}
\arguments{
\item{dir}{folder in which the files to be converted are stored}

\item{origin}{extension of the files to be converted, for example ("xlsx")}

\item{dest}{extension to which the files will be converted, for example ("csv").}

\item{recursive}{whether lo kook recursively in sub-folders, FALSE by default}
}
\value{
files are read/written with rio, the same name is kept
}
\description{
A function to convert most common formats. Specially useful when we need to compress to a certain format a lot of files.
}
\examples{
# convert all csv files in folder E:/test and sub-folders to parquet
convert_anything("E:/test",
                 "csv",
                 "parquet",
                 recursive = TRUE)

}
