% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_eurobase_codes.R
\name{convert_eurobase_codes}
\alias{convert_eurobase_codes}
\title{Convert Eurobase codes}
\usage{
convert_eurobase_codes(df)
}
\arguments{
\item{df}{the dataframe with the eurobase data}
}
\value{
a dataframe with new column names and codes
}
\description{
This function convert Eurobase column names and other codes (NACE) to the
codes used in transmissions. This allows an easier comparison between datasets.
}
\examples{
df<- load_sent_eurobase(folder="E:/data/REGACC/sent",
table_sel = "nama_10r_2coe",
country_sel = "ES",
time_min = "2022-01-01")

df1<-convert_eurobase_codes(df)
}
