% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut_plot.R
\name{donut_plot}
\alias{donut_plot}
\title{Create a donut plot for NQR or other purposes}
\usage{
donut_plot(value, colour = "#0E47CB", title)
}
\arguments{
\item{value}{the value to show inside the donut}

\item{colour}{the colour of the elemtent, by default EC blue}

\item{title}{Title of the chart}
}
\value{
a ggplot chart
}
\description{
Create a donut plot for NQR or other purposes
}
\examples{
donut_plot(value = 0.65, title = "show a percentage in a donut")

}
