% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurostat_find.R
\name{eurostat_find}
\alias{eurostat_find}
\title{Find Eurostat data sets by code, description or date of update}
\usage{
eurostat_find(sel_code, sel_desc, sel_update)
}
\arguments{
\item{sel_code}{a string ("nama_10r")}

\item{sel_desc}{a string ("GDP")}

\item{sel_update}{a date("2022-07-15")}
}
\value{
a data frame
}
\description{
This function looks at Eurostat data sets and finds those that match the full or
partial code or description provided. The user can filter by date of last update.
Arguments should be unique (only one string per argument).
}
\examples{
df<-eurostat_find(sel_code="nama_10r")
df<-eurostat_find(sel_code="nama_10r",sel_desc = "GDP")
df<-eurostat_find(sel_code="nama_10",sel_desc = "GDP", sel_update ="2022-07-15")

}
