% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_annual_NAMA}
\alias{get_annual_NAMA}
\title{Get a dataframe with the latest Eurobase annual NAMA}
\usage{
get_annual_NAMA(
  na_item_sel = "B1GQ",
  unit_sel = c("CP_MEUR", "CP_MPPS_EU27_2020"),
  min_time = "2000-01-01"
)
}
\arguments{
\item{na_item_sel}{which na_item to download, B1GQ by default}

\item{unit_sel}{which unit, CP_MEUR by and CP_MPPS_EU27_2020 by default}

\item{min_time}{minimum period to ge the data from, by default 2000-01-01}
}
\value{
a data.frame/data.table
}
\description{
This function downloads from Eurobase the latest NAMA data  and other variables in the table nama_10_gdp for the unit declared by the user.
}
\examples{
dt<- get_annual_NAMA(na_item_sel=c("B1GQ", "D1"), unit_sel= c("CP_MEUR","CP_MNAC"))
}
