% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_annual_NAMApc}
\alias{get_annual_NAMApc}
\title{Get a dataframe with the latest Eurobase annual NAMA per capita}
\usage{
get_annual_NAMApc(
  na_item_sel = "B1GQ",
  unit_sel = c("PC_EU27_2020_HAB_MEUR_CP", "PC_EU27_2020_HAB_MPPS_CP", "CP_EUR_HAB",
    "CP_PPS_EU27_2020_HAB"),
  min_time = "2000-01-01"
)
}
\arguments{
\item{na_item_sel}{which na_item to download, B1GQ by default}

\item{unit_sel}{which unit, CEUR and PPS are chosen by default}

\item{min_time}{minimum period to ge the data from, by default 2000-01-01}
}
\value{
a data.frame/data.table
}
\description{
This function downloads from Eurobase the latest NAMA data  and other variables in the table nama_10_gdp for the unit declared by the user.
}
\examples{
dt<- get_annual_NAMApc(na_item_sel=c("B1GQ","P31_S14"))
}
