% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denodo.R
\name{get_denodo}
\alias{get_denodo}
\title{Get a cleaned denodo extraction}
\usage{
get_denodo(T1001_1200 = FALSE)
}
\arguments{
\item{T1001_1200}{FALSE (default) extracts primary series from tables T1001, T1002, T1200 and T1300. TRUE extracts all series from T1001 and T1001_1200.}
}
\value{
a dataframe
}
\description{
The function connects to denodo and extracts either primary series (by default) or the complete A tables.
}
\examples{
partial<- get_denodo()
full<- get_denodo(full=TRUE)
}
