% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GDP.R
\name{get_regional_GDP}
\alias{get_regional_GDP}
\title{Get a dataframe with the latest Eurobase regional GDP}
\usage{
get_regional_GDP(
  unit_sel = c("MIO_EUR", "EUR_HAB", "EUR_HAB_EU27_2020", "MIO_NAC", "MIO_PPS_EU27_2020",
    "PPS_EU27_2020_HAB", "PPS_HAB_EU27_2020")
)
}
\arguments{
\item{unit_sel}{which unit, all by default. Options are "MIO_EUR","EUR_HAB","    EUR_HAB_EU27_2020", "MIO_NAC", "MIO_PPS_EU27_2020", "PPS_EU27_2020_HAB","    PPS_HAB_EU27_2020".}
}
\value{
a data.frame/data.table
}
\description{
This function downloads from Eurobase the latest regional GDP data in the table nama_10r_3gdp for the unit declared by the user.
}
\examples{
dt<- get_regional_GDP(unit_sel= c("MIO_EUR","EUR_HAB"))
}
