% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xml_basic_info.R
\name{get_xml_basic_info}
\alias{get_xml_basic_info}
\title{Get basic information from xml files}
\usage{
get_xml_basic_info(...)
}
\arguments{
\item{...}{same parameters as for load_xml}
}
\value{
four interactive html tables
}
\description{
This functions provides summary information (reporting periods, regions, flags and comments) from xml files. The output are dynamic HTML tables
}
\examples{
get_xml_basic_info(folder = "D:/data/REGACC/xml",
country_sel="AT",
time_min = "2021-10-01")

}
