% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{Load csv files}
\usage{
load_csv(
  folder = "//fame4prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE",
  country_sel,
  table_sel,
  sto_sel,
  unit_sel,
  time_min = "2021-10-01",
  time_max = "2099-01-01",
  consolidate = FALSE
)
}
\arguments{
\item{folder}{specifies the folder where the files are. By default is the server folder.}

\item{country_sel}{Country or countries to look for, by default all}

\item{table_sel}{table or tables to look for.}

\item{sto_sel}{NA item to look for ("B1G", c("B1G","EMP"))}

\item{unit_sel}{Unit to look for ("XDC", c("XDC","PC"))}

\item{time_min}{Date from where to look for.}

\item{time_max}{Date where to stop looking for.}

\item{consolidate}{TRUE to remove duplicated values, FALSE (default) to keep them all}
}
\value{
a data frame
}
\description{
This function loads csv files into R. The function will look by default on the server
folder for all countries and all tables but that can be changed by the user using
the arguments of the function. The unit multiplier is changed, if needed, to be
consistent with the one used by default elsewhere. Not all columns in the original file are kept
I selected the most useful ones for numerical analysis. For qualitative info (flags, comments) it
is recommended to use the related function regacc_load_xml.
}
\examples{
# Load all the files in the folder "//fame4prod.cc.cec.eu.int/fame-estat/econ/REGACC/INSPACE"
df<- load_csv()

# Load only table T1001.
df <- load_csv(table_sel = "T1001")

# Load all tables T1001 and T1300 for Slovenia and Luxembourg.
df <- load_csv(table_sel = c("T1001","T1300"),
                      country_sel = c("SI", "LU"))

# Load all files loaded between 2021-12-22 2022-01-02 and consolidate them.
df <- load_csv(time_min = "2021-12-22",
                      time_max = "2022-01-02",
                      consolidate = TRUE)

}
