% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_matis.R
\name{load_matis}
\alias{load_matis}
\title{Load Matis}
\usage{
load_matis(file, keep_cols)
}
\arguments{
\item{file}{The file to load.}

\item{keep_cols}{The columns to keep.}
}
\value{
a data frame/ data.table
}
\description{
Load excel extractions from REGACC for A_ tables (VAL.T.T1001.A.AT.W2.S1.S1.B.B1G._T.B.V.N.XDC) extracted horizontally and select the columns to keep.

By default it keeps the most useful columns. To keep all columns use: keep_cols= c("VAL","type", "table_identifier", "freq", "ref_area",
"counterpart_area","ref_sector","counterpart_sector","accounting_entry", "sto", "activity", "valuation", "prices",
"transformation","unit_measure", "time","values","flag) or any subset.
}
\examples{
df<-load_matis("D:/03_Regional Accounts/03D_Data Production/2021/BE/matis/t1200_BE_01.xlsx",keep_cols= c("type","ref_area","sto","activity","time_period","obs_value"))


}
