% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_matis_txt.R
\name{load_matis_txt}
\alias{load_matis_txt}
\title{Load Matis big extractions}
\usage{
load_matis_txt(file, keep_cols)
}
\arguments{
\item{file}{The file to load.}

\item{keep_cols}{The columns to keep.}
}
\value{
a data frame/ data.table
}
\description{
Load txtextractions from REGACC for A_ tables (VAL.V.T1001.A.AT.W2.S1.S1.B.B1G._T.B.V.N.XDC) extracted horizontally and select the columns to keep.

By default it keeps the most useful columns. To keep all columns use: keep_cols= c("VAL","type", "table_identifier", "freq", "ref_area",
"counterpart_area","ref_sector","counterpart_sector","accounting_entry", "sto", "activity", "valuation", "prices",
"transformation","unit_measure", "prices", "time","values","flag) or any subset.
}
\examples{
df<-load_matis_txt("E:/regacc/data/matis/extract_data_manager_39_COUNTRIES_230106_120132.txt")


}
