% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sent_eurobase.R
\name{load_sent_eurobase}
\alias{load_sent_eurobase}
\title{Load files sent to Eurobase}
\usage{
load_sent_eurobase(
  folder,
  table_sel,
  country_sel,
  time_min = "2019-01-01",
  time_max = "2023-07-11",
  consolidate = FALSE
)
}
\arguments{
\item{folder}{specifies the folder where the files are (zip files)}

\item{table_sel}{the Eurobase table to look for. It only accepts a single argument (nama_10r_2gdp, nama_10r_3gdp, nama_10r_3popgdp, nama_10r_3gva, nama_10r_3empers, nama_10r_2coe, nama_10r_2gfcf, nama_10r_2emhrw, nama_10r_2hhinc, nama_10r_2gvavr,nama_10r_2lp10, nama_10r_2nlp, nama_10r_3nlp)}

\item{country_sel}{country or countries to look for}

\item{time_min}{date of publication from where to start looking (yyyy-mm-dd)}

\item{time_max}{date of publication from where to stop looking (yyyy-mm-dd)}

\item{consolidate}{TRUE to remove duplicated values, FALSE (default) to keep them all}
}
\value{
a data frame
}
\description{
This function looks at the folder where the files sent to Eurobase are stored, loads them and puts them together in a data frame.
}
\examples{
df<- load_sent_eurobase(folder = "E:/data/REGACC/sent/source",
table_sel="nama_10r_2gvagr",
country_sel = c("ES","PT"),
time_min = "2020-01-01",
time_max = "2021-03-16",
consolidate = TRUE)
}
