% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_use.R
\name{mem_use}
\alias{mem_use}
\title{A function that returns a data.frame with the memory used by different objects}
\usage{
mem_use(sort = "size", decreasing = TRUE, limit = 10)
}
\arguments{
\item{decreasing}{decreasing}

\item{pos}{position, left always to 1}

\item{pattern}{pattern to filter}

\item{order.by}{order by default by size}

\item{head}{head}

\item{n}{number of objects in the ouput 10 by default}
}
\value{
a data.frame
}
\description{
A function that returns a data.frame with the memory used by different objects
}
\examples{

mem_use()

}
