% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_colh.R
\name{ra_colh}
\alias{ra_colh}
\title{Create standard horizontal bar charts}
\usage{
ra_colh(
  dat,
  hor,
  ver,
  clr,
  leg = TRUE,
  fac,
  wdt = 0.98,
  hbr = 5,
  acc = 1,
  bar = "stack"
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{clr}{the variable mapped to the fill aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{wdt}{width of the bar from 0 to 1, default 0.98}

\item{hbr}{parameter for the number of horizontal breaks, input for scales::break_extended}

\item{acc}{number of decimals for X axis (0.1 means one decimal)}

\item{bar}{type of bar, "stack" by default, other options "dodge" and "fill"}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard scatter charts for different purposes
}
\examples{
df<- dataregacc::NQR \%>\%
 filter(country =="SK" & vintage =="2021" & NUTS =="2") \%>\%
 pivot_wider(names_from=sto,
             values_from=obs_value)

ra_colh(dat=df \%>\% filter(time_period==2020),
           hor =EMP,
           ver = ref_area,
           clr = ref_area,
           wdt = 1)
}
