% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_heatmap.R
\name{ra_heatmap}
\alias{ra_heatmap}
\title{Create standard barcharts}
\usage{
ra_heatmap(
  dat,
  hor,
  ver,
  clr,
  leg = TRUE,
  fac,
  lwt = 0.5,
  lcl = "white",
  pal = "monoblue"
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{clr}{the variable mapped to the fill aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{lwt}{linewdth for the squares}

\item{lcl}{line colour for the squares}

\item{pal}{fill palette to be chosen from scale_fill_regacc_c}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard scatter charts for different purposes
}
\examples{
df<- dataregacc::NQR \%>\%
 filter(country =="SK" & vintage =="2021" & NUTS =="2") \%>\%
 pivot_wider(names_from=sto,
             values_from=obs_value)

ra_heatmap(dat=df \%>\% drop_na(D1),
              hor=time_period,
              ver= ref_area,
              clr=D1)
}
