% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_line.R
\name{ra_line}
\alias{ra_line}
\title{Create standard line charts}
\usage{
ra_line(
  dat,
  hor,
  ver,
  grp,
  clr,
  leg = TRUE,
  fac,
  lsz = 0.8,
  hbr = 3,
  vbr = 4,
  acc = 1,
  exm = 0,
  exa = 0.2
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{grp}{the variable to use for grouping the data}

\item{clr}{the variable mapped to the colour aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{lsz}{line size}

\item{hbr}{parameter for the number of horizontal breaks, input for scales::break_extended}

\item{vbr}{parameter for the number of vertical breaks, input for scales::break_extended}

\item{acc}{number of decimals for Y axis (0.1 means one decimal)}

\item{exm}{multiplicative parameter for the scale limits}

\item{exa}{additive parameter for the scale limits}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard line charts for different purposes
}
\examples{
df<- dataregacc::NQR \%>\%
 filter(country =="SK" & vintage =="2021" & NUTS =="2")
ra_line(dat=df,
                 hor=time_period,
                 ver=obs_value,
                 grp=ref_area,
                 hbr=3,
                 vbr=4,
                 clr=ref_area,
                 leg=FALSE,
                 fac=sto)
}
