% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_logo.R
\name{ra_logo}
\alias{ra_logo}
\title{adding new Eurostat logo to charts}
\usage{
ra_logo(chart, colour = "blue", left = 0.87, bottom = 0, right = 1, top = 0.1)
}
\arguments{
\item{chart}{ggplot chart in which the logo has to be inserted}

\item{colour}{"blue" by default or "white"}

\item{left}{left alignment , default 0.9}

\item{bottom}{bottom alignment , default 0.0}

\item{right}{right alignment , default 0.99}

\item{top}{top alignment , default 0.1}
}
\value{
a ggplot
}
\description{
adding new Eurostat logo to charts
}
\examples{
p<-ggplot(dataregacc::regacc_eurobase \%>\%
         filter(country=="EL" &
                NUTS==1 &
                unit_measure=="EUR_HAB" &
                table=="gdp2"),
  aes(time_period,obs_value,colour=ref_area))+
  geom_line()+
  theme_ra()+
  scale_colour_ra()+
  labs(title="Greece NUTS 1 regions GDP",
       subtitle="EUR per inhabitant")

q<-ra_logo(p)
q
}
