% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_palettes.R
\name{ra_palettes}
\alias{ra_palettes}
\title{Function to create palettes from colours}
\usage{
ra_palettes(
  name,
  n,
  all_palettes = ra_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{name of the palette ("light","dark","primary","monoblue"...)}

\item{n}{number of colours}

\item{all_palettes}{list of all palettes}

\item{type}{continue or discrete}
}
\value{
a ready to use palette
}
\description{
Function to create palettes from colours
}
\examples{
scales::show_col(ra_palettes(name="light"))
}
