% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra_scatter.R
\name{ra_scatter}
\alias{ra_scatter}
\title{Create standard scatter charts}
\usage{
ra_scatter(
  dat,
  hor,
  ver,
  clr,
  leg = TRUE,
  fac,
  psz = 3,
  hbr = 4,
  vbr = 4,
  acch = 1,
  accv = 1
)
}
\arguments{
\item{dat}{the data frame with the data}

\item{hor}{the name of the variable in the horizontal axis}

\item{ver}{the name of the variable in the vertical axis}

\item{clr}{the variable mapped to the colour aesthetic}

\item{leg}{set to FALSE if you do not want to show the legend}

\item{fac}{the (unique) variable to use if small multiples are wanted}

\item{psz}{point size}

\item{hbr}{parameter for the number of horizontal breaks, input for scales::break_extended}

\item{vbr}{parameter for the number of vertical breaks, input for scales::break_extended}

\item{acch}{number of decimals for Y axis (0.1 means one decimal)}

\item{accv}{number of decimals for Y axis (0.1 means one decimal)}
}
\value{
a ggplot
}
\description{
This functions makes easy to create standard scatter charts for different purposes
}
\examples{
df<- regacc::NQR \%>\%
 filter(country =="SK" & vintage =="2021" & NUTS =="2") \%>\%
 pivot_wider(names_from=sto,
             values_from=obs_value)
ra_scatter(dat=df,
          hor=B1G,
          ver=B6N,
          clr=ref_area)
}
