% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regacc-package.R
\docType{package}
\name{regacc-package}
\alias{regacc}
\alias{regacc-package}
\title{regacc: Set of functions used in regacc}
\description{
Collection of functions for Regional Accounts to be used together with {dataregacc} and other handy functions.
}
\keyword{internal}
