% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_completeness.R
\name{report_completeness}
\alias{report_completeness}
\title{Report completeness as in NQR}
\usage{
report_completeness(df, table_sel, verbose = FALSE)
}
\arguments{
\item{table_sel}{individual table to check}

\item{verbose}{FALSE(default) shows the result as a donut chart. TRUE to get a table with the details.}

\item{data}{a data frame with the reported data}
}
\value{
completeness ratios and missing series
}
\description{
The function provides the same ratios as reported in the NQR and an interactive table with the missing values.
}
\examples{
df<- load_xml("D:/regacc/data/input/xml",
country_sel = "ES",
table_sel = c("T1001","T1002","T1200","T1300"),
time_min = "2021-10-01")

report_completeness (df,"T1001")

}
