% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_NQR_revision.R
\name{report_NQR_revision}
\alias{report_NQR_revision}
\title{NQR revision file
Function to create the NRQ revision report that we sent to countries
Needs to be updated every year to change some parameters}
\usage{
report_NQR_revision(dat, country_sel, output_dir)
}
\arguments{
\item{dat}{external csv file with the data}

\item{country_sel}{country}

\item{output_dir}{folder to store the generated excel file}
}
\value{
an excel report
}
\description{
NQR revision file
Function to create the NRQ revision report that we sent to countries
Needs to be updated every year to change some parameters
}
\examples{
report_NQR_revision(dat="D:/DK_new.csv",
                    country_sel = "DK",
                    output_dir="D:/")

}
