% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_transmission.R
\name{report_transmission}
\alias{report_transmission}
\title{Report transmissions}
\usage{
report_transmission(
  folder,
  country_sel,
  table_sel,
  time_min = "2021-11-01",
  time_max = "2099-10-01"
)
}
\arguments{
\item{folder}{folder where files area.}

\item{country_sel}{optionally specify a country.}

\item{table_sel}{optionally specify a table.}

\item{time_min}{optionally specify a minimum date.}

\item{time_max}{optionally specify a maximum date}
}
\value{
a data frame
}
\description{
The function looks at a specific folder with the transmitted data from countries and provides a table with the dates
}
\examples{
report_transmission("E:/regacc/data/input/xml")
report_transmission("E:/regacc/data/input/xml", country_sel="SI", table_sel="T1300")
}
