% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_ra.R
\name{scale_colour_ra}
\alias{scale_colour_ra}
\title{colour palette for charts}
\usage{
scale_colour_ra(name = "light")
}
\arguments{
\item{name}{name of the palette. "full" by default. Other options ("primary","monoblue","monofuchsia") as defined in regacc_palettes()}
}
\value{
a colour palette
}
\description{
colour palette for charts
}
\examples{
library(dataregacc)
library(tidyverse)
ro_d61<- eurobase \%>\%
 filter(sto == "D61" &
        accounting_entry == "D" &
        unit_measure == "MIO_NAC" &
        country =="RO")

ggplot(ro_d61 \%>\% filter(NUTS==2),aes(time_period,obs_value,colour=ref_area))+
 geom_line(size=1)+
 scale_colour_ra()
}
