% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_ra_c.R
\name{scale_fill_ra_c}
\alias{scale_fill_ra_c}
\title{fill palette for charts}
\usage{
scale_fill_ra_c(name = "monoblue")
}
\arguments{
\item{name}{name of the palette. "full" by default. Other options ("primary","monoblue","monofuchsia")}
}
\value{
a continuous fill palette
}
\description{
fill palette for charts
}
\examples{
a fill palette
library(dataregacc)
library(tidyverse)
ro_d61<- eurobase \%>\%
 filter(sto == "D61" &
        accounting_entry == "D" &
        unit_measure == "MIO_NAC" &
        country =="RO")

ggplot(ro_d61,aes(time_period,ref_area,fill=obs_value))+
geom_tile()+
scale_fill_ra_c()
}
