% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_DT.R
\name{show_DT}
\alias{show_DT}
\title{Create an HTML interactive table with DT}
\usage{
show_DT(x, caption = "")
}
\arguments{
\item{x}{a data frame}
}
\value{
DT table
}
\description{
An interactive HTML table with some fancy options (filter,sort,export to excel) is generated from a data frame.
}
\examples{

show_DT(dataregacc::NUTS_2021)

}
