% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_excel.R
\name{show_in_excel}
\alias{show_in_excel}
\title{Export to excel a data frame}
\usage{
show_in_excel(.data)
}
\arguments{
\item{.data}{}
}
\value{
an excel file in a temporary folder
}
\description{
For closer inspection of small data frames this functions exports it to excel and opens the file.
}
\examples{

show_in_excel(dataregacc::NUTS_2021)
}
