% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ra.R
\name{theme_ra}
\alias{theme_ra}
\title{Default theme options for regacc charts}
\usage{
theme_ra(font = "MS Sans Serif")
}
\arguments{
\item{font}{Font to be used by default MS Sans Sheriff}
}
\value{
a set of theme paremeters
}
\description{
Default theme options for regacc charts
}
\examples{
ggplot(dataregacc::eurobase \%>\%
         filter(country=="EL" &
                NUTS==1 &
                unit_measure=="EUR_HAB" &
                table=="gdp2"),
  aes(time_period,obs_value,colour=ref_area))+
  geom_line()+
  theme_ra()+
  scale_colour_ra()+
  labs(title="Greece NUTS 1 regions GDP",
       subtitle="EUR per inhabitant")
}
