% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denodo.R
\name{get_denodo}
\alias{get_denodo}
\title{Get a cleaned denodo extraction}
\usage{
get_denodo(extraction)
}
\arguments{
\item{extraction}{view to select from Denodo (hv_fame_regacc4regacc_all_dbs,"hv_fame_regacc4regacc_all_dbs","hv_fame_regacc4regacc_t1001","hv_fame_regacc4regacc_t1002","hv_fame_regacc4regacc_t1200", "hv_fame_regacc4regacc_t1300","hv_fame_regacc4regacc_t1001_1200", "hv_fame_regacc4regacc_tnama" )}
}
\value{
a dataframe
}
\description{
The function connects to denodo and extracts either primary series (by default) or the complete A tables.
}
\examples{
tnama<-get_denodo_new(extraction = "hv_fame_regacc4regacc_tnama")

}
