% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denodo.R
\name{get_denodo}
\alias{get_denodo}
\title{Get a cleaned denodo extraction}
\usage{
get_denodo(onlyV = FALSE)
}
\arguments{
\item{onlyV}{FALSE (default) extracts primary T and V series from tables T1001, T1002, T1200 and T1300. This is used for creating country csv files for the analysis. TRUE extracts V series for several units. The file is used for other purposes}
}
\value{
a dataframe
}
\description{
The function connects to denodo and extracts either primary series (by default) or the complete A tables.
}
\examples{
partial<- get_denodo()
full<- get_denodo(full=TRUE)
}
