% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denodo.R
\name{get_denodo}
\alias{get_denodo}
\title{Get a cleaned denodo extraction}
\usage{
get_denodo(full = FALSE)
}
\arguments{
\item{full}{FALSE (default) just extracts primare series. TRUE all series}
}
\value{
a dataframe
}
\description{
The function connects to denodo and extracts either primary series (by default) or the complete A tables.
}
\examples{
partial<- get_denodo()
full<- get_denodo(full=TRUE)
}
