% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_NQR_data.R
\name{update_NQR_data}
\alias{update_NQR_data}
\title{Update NQR data set with 2022 transmissions}
\usage{
update_NQR_data(
  country_sel,
  input_file = "data/denodo/all_primary.parquet",
  output_dir = "NQR/data"
)
}
\arguments{
\item{country_sel}{country to update}

\item{input_file}{parquet file with an extraction of primary data}

\item{output_dir}{output folder for the file}
}
\value{
a csv file
}
\description{
This function puts together the existing NQR vintages (up to 2021) and the newest one
from the validated series from a denodo extraction and creates a csv file
with the results in the chosen folder
}
\examples{
update_NQR_data(country_sel = "DK",
input_file = "D:/check_input/data/denodo/all_primary.parquet",
output_dir= "D:/")
}
